fix tests for ES metrics
